Use state to store club token minted
