+++ src/components/TokenMinting.tsx
@@ -1,13 +1,18 @@
 import { ThirdwebSDK } from "@thirdweb-dev/sdk";
 import React from "react";
+import { getAppControllers } from "../controllers";
 import { ThirdWebController } from "../controllers/thirdweb";
 
 export interface ITokenMintingStates {
-  sdkController: ThirdWebController | any;
+  sdkController: ThirdWebController;
+  sdk: ThirdwebSDK;
+  dropTokenAddress: string;
 }
 
 export const INITIAL_STATE = {
-  sdkController: ""
+  sdkController: getAppControllers().thirdweb,
+  sdk: getAppControllers().thirdweb.sdk,
+  dropTokenAddress: "",
 };
 export class TokenMinting extends React.Component {
   public state: ITokenMintingStates;
@@ -27,12 +32,18 @@ export class TokenMinting extends React.Component {
     sdkController.getSdkAddress();
   }
 
+  public async deployClubTokenContract(name_input: string, symbol_input: string) {
+    const { sdkController } = this.state;
+    const dropTokenAddress = await sdkController.getClubTokenAddress(name_input, symbol_input);
+    this.setState({dropTokenAddress});
+  }
+
   render() {
     let name_input: string, symbol_input: string;
-    const { sdkController } = this.state;
+    const { sdkController, dropTokenAddress} = this.state;
     return (
       <>
-        {!sdkController.dropTokenAddress ? (
+        {!dropTokenAddress ? (
           <>
             <h4>Token minting section</h4>
             <div>
@@ -55,14 +66,14 @@ export class TokenMinting extends React.Component {
             </div>
             <button
               onClick={() =>
-                sdkController.getClubTokenAddress(name_input, symbol_input)
+                this.deployClubTokenContract(name_input,symbol_input)
               }
             >
               Mint drop token
             </button>
           </>
         ) : (
-          <></>
+          <>Token minted! Contract address: {dropTokenAddress}</>
         )}
       </>
     );
